refactor safe utils to utils
